+++ 1.x/Source/dgopher.cc
@@ -91,7 +91,9 @@ qint64 dgopher::readData(char *data, qint64 maxSize)
   if(number < 0)
     return -1;
 
-  memcpy(data, m_html.constData() + m_offset, number); // number >= 0
+  memcpy(data,
+	 m_html.constData() + m_offset,
+	 static_cast<size_t> (number)); // number >= 0
   m_offset += number;
   return number;
 }
@@ -121,7 +123,8 @@ void dgopher::download(void)
 	  SIGNAL(readyRead(void)),
 	  this,
 	  SLOT(slotReadyReadForDownload(void)));
-  m_socket->connectToHost(url().host(), url().port());
+  m_socket->connectToHost
+    (url().host(), static_cast<quint16> (url().port()));
 }
 
 void dgopher::initialize(void)
@@ -156,7 +159,8 @@ void dgopher::load(void)
 	  SIGNAL(readyRead(void)),
 	  this,
 	  SLOT(slotReadyRead(void)));
-  m_socket->connectToHost(url().host(), url().port());
+  m_socket->connectToHost
+    (url().host(), static_cast<quint16> (url().port()));
 }
 
 void dgopher::slotConnected(void)
