+++ 1.x/Source/dnetworkaccessmanager.cc
@@ -155,7 +155,7 @@ dnetworkerrorreply::dnetworkerrorreply
 
   request.setAttribute(QNetworkRequest::CacheSaveControlAttribute, false);
   setRequest(request);
-  url.setScheme(scheme.mid(qstrlen("dooble-")));
+  url.setScheme(scheme.mid(static_cast<int> (qstrlen("dooble-"))));
   m_content.append("<html>");
   m_content.append("<head>");
   m_content.append(QString("<title>%1</title>").
@@ -371,7 +371,7 @@ dnetworksslerrorreply::dnetworksslerrorreply
 
   request.setAttribute(QNetworkRequest::CacheSaveControlAttribute, false);
   setRequest(request);
-  url.setScheme(scheme.mid(qstrlen("dooble-ssl-")));
+  url.setScheme(scheme.mid(static_cast<int> (qstrlen("dooble-ssl-"))));
   m_content.append("<html>");
   m_content.append("<head>");
   m_content.append(QString("<title>%1</title>").
